Update to 5.22.0 and switch over to prebuilt binary.
Unfortunately Signal Desktop does not support nodejs 16 LTS, as it uses an incredibly out-of-date version of better-sqlite3 that attempts to use non-existant APIs, a version of node-sass that doesn't support 16.x, and a version of sass-loader from the stone ages. As a result, switch to its Debian prebuilt binary to allow for continued use of Signal.